refactor

+++ dwm.c
@@ -1633,7 +1633,7 @@ monitor_draw_bar(Monitor *monitor) {
             if (client) {
                 snprintf(tagdisp, TAGWIDTH, "%s", tags[i]);
             } else {
-                int n = strcspn(masterclientontag[i], tag_label_delim);
+                ulong n = strcspn(masterclientontag[i], tag_label_delim);
                 masterclientontag[i][n] = '\0';
                 snprintf(tagdisp, TAGWIDTH, tag_label_format, tags[i], masterclientontag[i]);
             }
